how to get a handle on the hub provider?
