+++ relayer/processor/path_end_runtime.go
@@ -2,6 +2,7 @@ package processor
 
 import (
 	"context"
+	"errors"
 	"fmt"
 	"sync"
 	"time"
@@ -288,16 +289,39 @@ func (pathEnd *pathEndRuntime) handleCallbacks(c IBCMessagesCache) {
 	}
 }
 
-type DymRollappChan
+func SendGenesisTransfer(
+	ctx context.Context,
+	hubC provider.ChainProvider,
+	raC provider.ChainProvider,
+) error {
+	hub, ok := hubC.(provider.DymensionHubProvider)
+	if !ok {
+		return errors.New("not dymension hub provider")
+	}
+
+	channelID, err := hub.GetCanonicalChan(ctx, raC.ChainId())
+	if err != nil {
+		return fmt.Errorf("get canonical chan: %w", err)
+	}
+
+	ra, ok := raC.(provider.RollappProvider)
+	if !ok {
+		return errors.New("not rollapp provider")
+	}
+	return ra.TrySendGenesisTransfer(ctx, channelID)
+}
 
 // DYMENSION
 func (pathEnd *pathEndRuntime) handleDymensionCallbacks(ctx context.Context, c IBCMessagesCache) {
-	cache, ok := c.ChannelHandshake[chantypes.EventTypeChannelOpenConfirm]
+	_, ok := c.ChannelHandshake[chantypes.EventTypeChannelOpenConfirm]
 	if !ok {
 		return
 	}
 	var hub *relayer.Chain
-	err := relayer.SendGenesisTransfer(ctx, hub, pathEnd.chainProvider)
+	err := SendGenesisTransfer(ctx, hub, pathEnd.chainProvider)
+	if err != nil {
+		pathEnd.log.Error("Send rollapp genesis transfer to hub. Operator should retry using CLI.", zap.Error(err))
+	}
 }
 
 func (pathEnd *pathEndRuntime) shouldTerminate(ibcMessagesCache IBCMessagesCache, messageLifecycle MessageLifecycle) bool {
@@ -522,7 +546,9 @@ func (pathEnd *pathEndRuntime) mergeCacheData(
 
 	pathEnd.handleCallbacks(d.IBCMessagesCache)
 
-	pathEnd.handleDymensionCallbacks(ctx, d.IBCMessagesCache)
+	if pathEnd.chainProvider.IsDymensionRollapp() {
+		pathEnd.handleDymensionCallbacks(ctx, d.IBCMessagesCache)
+	}
 
 	if pathEnd.shouldTerminate(d.IBCMessagesCache, messageLifecycle) || terminate {
 		cancel()
